refactor(agents): update bookmark terminology and streamline bookmark functionality in AgentCard component
